requirements.txt
# local package
-e .

# external requirements
boto3
joblib
scikit-learn
pandas>=2.0.0
numpy
matplotlib
seaborn
nltk
spacy
textblob
fastapi
uvicorn
mlflow
dvc
dagshub
pyyaml
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
pathlib2
rsa>=4.7